Update UNet_Model
he DoubleConv class defines a block that consists of two convolutional layers each followed by batch normalization and ReLU activation. The UNet class uses multiple instances of the DoubleConv block to build the encoder (downsampling) and decoder (upsampling) parts of the UNet architecture. The in_channels argument specifies the number of input channels, out_channels specifies the number of output channels (which is set to 1 for binary segmentation), and features specifies the number of filters/channels for each layer in the encoder and decoder.

In the forward pass, the input tensor is first passed through the encoder, which downsamples the spatial resolution of the feature maps while increasing the number of channels. The output of the last encoder layer (bottleneck) is then passed to the decoder, which upsamples the feature maps back to the original resolution while reducing the number of channels. Skip connections from the encoder are also concatenated with the feature maps in the decoder to help preserve spatial information. Finally, the output tensor is passed through a 1x1 convolutional layer to produce the final segmentation map.